refactor: split `ErrorCode` and `SuccessCode` enums into separate interfaces and enums for better modularity and maintainability
