+++ src/main/java/com/yeoljeong/tripmate/exception/constants/CommonErrorCode.java
@@ -1,0 +1,26 @@
+package com.yeoljeong.tripmate.exception.constants;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum CommonErrorCode implements ErrorCode {
+	INVALID_INPUT(HttpStatus.BAD_REQUEST, "잘못된 입력입니다."),
+	UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "인증이 필요합니다."),
+	FORBIDDEN(HttpStatus.FORBIDDEN, "접근 권한이 없습니다."),
+	NOT_FOUND(HttpStatus.NOT_FOUND, "리소스를 찾을 수 없습니다."),
+	INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 오류가 발생했습니다."),
+	METHOD_NOT_ALLOWED(HttpStatus.METHOD_NOT_ALLOWED, "허용되지 않는 메서드입니다.")
+	;
+
+	private final HttpStatus status;
+	private final String message;
+
+	@Override
+	public int getCode() {
+		return this.status.value();
+	}
+}
+
